+++ CHANGELOG.md
@@ -8,8 +8,8 @@
 - Dropping undocumented `#.upgradeToMixed` & `#.upgradeToMulti`.
 - Refactoring internal indices.
 - Fixing edge iteration wrt. self loops.
-- Improving performance of edge iterators.
-- Improving perfromance of neighbor iterarors.
+- Improving performance of edge iteration.
+- Improving perfromance of neighbor iteration.
 - Improving performance of node deletion.
 - Improving performance of internal edge key generator.
 - Improving memory usage.